use thrift datastructures to serialize errors
Summary:
For some reason the async requests errors were stored in blobstore as strings.
We already have datastructures to properly store them without information loss:
the actual thrfit exceptions that are going the be returned to the end user, we
can't use them directly but we can create structs that with the same structure
so the casting between them will be trivial.

This diff changes what we store. Unfortunately we still deserialize to
MononokeError and serialize back to thrift on pooling. The rest of this diff
stack will deal with that.

Reviewed By: StanislavGlebik

Differential Revision: D28747623

fbshipit-source-id: 7506f1a29220b49482f40d1db8e6b552dabd679a