mononoke: check if write is allowed before calling megarepo methods
Summary:
All megarepo methods write to a repository, so we need to check if write is
allowed to a given repo, and previously we weren't checking that.
Let's fix it and Let's start doing so now by trying to get RepoWriteContext
for the target repo. If writes are not allowed then RepoWriteContext fails.

Reviewed By: farnz

Differential Revision: D28838994

fbshipit-source-id: e45d4fe72603e7fe2755141874fc4125998bfed8