mononoke: fix validation failure when copyfile is used
Summary:
Turned out we still don't process copyfile attribute correctly. copyfile
creates two overrides like
{
  "copyfile source" -> default prefix + "/copyfile source"
  "copyfile source" -> "destination"
}

However we also have a check that validates that there no paths (that includes
default path, overrides, linkfiles etc) that are prefixes of each other. And
this check fails in the copyfile case, even though the path maps to exactly the
same path as if default prefix was applied.

Let's fix it. Also note that we haven't seen this failure in the integration
test because we didn't run verify_paths in tests. This diff fixes it as well.

Reviewed By: mitrandir77

Differential Revision: D28992456

fbshipit-source-id: 5fd993914b189cf768ba03010194b1c26026f7a8