mononoke/blobstore: add mysql sync queue latency logging
Summary:
This diff adds logging for the sync queue writes latency to the Mononoke Blobstore Trace Scuba table and creates a new column `Queue`, which can be used later on for the WAL logging purposes too.

This is needed to better understand out current situation with the MySQL base queue and compare it with the other solutions.

Reviewed By: ahornby

Differential Revision: D29015723

fbshipit-source-id: a2d7ad17101bb456ceae9060c39bfecb06644326