eden: move declare hgPath, hgPythonPath in HgImporter.h
Summary:
I'd like to be able to override this from unit tests. The underlying goal is
to let us make this path absolute in unit tests, so that we can make Eden unit
tests that rely on hgPath work on Buck v2, where paths are relative.

xavierd suggested overriding the glfag would be the easiest course of action
so that's what this does.

Reviewed By: xavierd

Differential Revision: D29060827

fbshipit-source-id: c4084072e3eef9cc5c9d404d10b3e19028f177ea