mononoke_types: make repo id suffix pattern a separate constant
Summary:
The `REPO_PREFIX_REGEX` attempts to match on the `.` of the repo id suffix,
however since this is a regex metacharacter, it needs to escape it.  Currently
it does this in the regex itself, but this is fragile.

Instead, let's have a separate constant for "the pattern that matches the repo
suffix", which has the value `"\."`.

Reviewed By: liubov-dmitrieva

Differential Revision: D29067723

fbshipit-source-id: 1e1a0a3ecf2b02d4e44fad1eaab57804f3dd020c