New `manual_scrub` mode for checking that a write-mostly store is populated
Summary:
I'm seeing significant Zippy load when I do a check scrub of our big repo to make sure that it's all in SQL Blobstore as well as our main blob stores.

Teach scrub to not bother talking to the main blobstores unless the write-mostly blobstore is either missing the data or unable to retrieve it.

Reviewed By: ahornby

Differential Revision: D29233349

fbshipit-source-id: 1127129ff283477558cddb03686c3c13aee47fb5