mononoke: no need to revisit deferred nodes in OldestFirst mode
Summary:
In the walker, an Option<NodeData> value of None is used to indicate that no data could be found for a node, and that for derived data mappings we should try again to load it later, when it may have been derived.

When a node is outside the chunk boundary this isn't appropriate,  we should just mark as visited and move on, which is what this change does.

Reviewed By: farnz

Differential Revision: D29230223

fbshipit-source-id: c2afdee9b914af89c7954c8e6a7d17a174df7ed1