packblob: generalise key prefixes
Summary:
Packblob currently expects key prefixes of the form `repoNNNN.` to be stripped , but also allows keys without this prefix. For the ephemeral blobstore we want to allow prefixes of the form `ephXXX.repoNNNN.` as well.

Generalise packblob so that we can have multiple key prefixes.

Packblob will enforce that none of the blobs in the packblob have a prefix that matches any of the patterns - this will prevent us from accidentally storing `repoNNNN.`-prefixed blobs in an ephemeral blobstore that requires `ephXXX.repoNNNN.` prefixes, for example.

Reviewed By: liubov-dmitrieva

Differential Revision: D29067720

fbshipit-source-id: 953909d47c9c4af91b529bcc684340d26411463d