ephemeral_blobstore: add ephemeral blobstore
Summary:
Implement get and put for the ephemeral blobstore.  This allows blobs to
be stored and retrieved in bubbles.

Ephemeral bubbles always have a repo associated with them when they are opened,
to simplify blob prefixing.  It is valid for a bubble id to have multiple repos
associated with it, but they must be accessed separately, and in practice this
won't be used.

Reviewed By: StanislavGlebik

Differential Revision: D29067722

fbshipit-source-id: d870f695fc1d0c825fdaec9337c82a13209165ce