Add git LFS support to gitimport and grepo branch_forest.
Summary:
Add git LFS support to gitimport and grepo branch_forest.

I did not want to add the parsing of .gitattributes and .lfsconfig to the gitimport library. This needs to be done by the users of gitimport before the import is started, And the GitImportLfs object needs to be configured accordingly. Currently we are extrating this data from the manifest files for the "g"repo imports.

I am not sure the simple git-lfs download client works with other git-lfs server back ends then Dewey. But it is a fairly simple implementation and it should be easy to extend to be more generic.

Reviewed By: farnz

Differential Revision: D29082867

fbshipit-source-id: a7b0272147b3d44a0b6b9782d2a1b8ec94653b8f