Add type RedactionKeyList
Summary:
This adds the blob object RedactionKeyList, which just contains a list of Strings, each of which will be a key to be redacted.

This will be stored on the blobstore, while a key to this object will be stored in configerator.

Some stuff that might be worth discussing:
- This class just holds a list of strings, per se it doesn't have much to do with redaction. If we want to change this to a more generic object like `KeyList`, I'm happy to do it. By default I'll leave it like this.
- I used serde (more precisely, json) to (de)serialise it. The only reason I did it was because I wanted to make this as simple as possible, from what I see in other objects need to define a thrift struct with the same config, then write `into/from_thrift` implementations. If preferred, I can do that.

It's not used in this diff, will be used in the future, I split it mostly to make it easier to review.

Reviewed By: markbt

Differential Revision: D29033597

fbshipit-source-id: 5550dbf58c5214201b739f8150fd06471bd67ab8