mononoke_api: detect multiple copies or renames when diffing
Summary:
When diffing a changeset with its parents, if a file is copied to multiple places, then we should include all of those copies in the diff.

Furthermore, if the file is also removed, then the *first* of those copies
should be considered a move.  Note that "first" here means the first in the
lexicographic ordering of the repository manifest.

Reviewed By: liubov-dmitrieva

Differential Revision: D29359516

fbshipit-source-id: eeed630c2e4d20f3fb8c923611a0433c74fd25d0