mutation: make obsolete() non-lazy
Summary:
Previously, the obsolete() set (from repo.dageval) is a "meta" (lazy) set that
might trigger frequent remote lookups. Calculate the obsolete() set into a
static set form so it does not trigger remote lookups.

This avoids expensive remote lookups in:

  File "/data/users/quark/fbsource/fbcode/een/scm/edenscm/hgext/smartlog.py", line 432, in <lambda>
    hidenodes = repo.dageval(lambda: obsnodes - heads(obsnodes) - roots(obsnodes))d

This issue was found out by setting env vars:

  EDENSCM_LOG=dag::protocol=debug

to find the commit hashes, and:

  EDENSCM_DISABLE_REMOTE_RESOLVE=a8209cbe13bcdc42d4999ffaa15dc9881660d5d,76d91d23d583f92b6c909b6747db1b75229d5be5

to find the traceback.

Reviewed By: andll

Differential Revision: D29404055

fbshipit-source-id: cc8bdd5347b38c1681565455874ed22f1a87321e