localrepo: disable EdenAPI for non-Mononoke repos
Summary: Previously, hg would always try to fetch data via EdenAPI, even when doing so wouldn't make sense (for example, a local test repo). The only time it makes sense to try to use EdenAPI at all is when the repo in question is supported by Mononoke. As a quick heuristic, let's disable EdenAPI if `paths.default` is not set to a `mononoke://` URL.

Differential Revision: D29372739

fbshipit-source-id: b3601e529ed44c355cb192aedf4332317f4e3132