mononoke: add --logview-additional-level-filter
Summary:
While debugging T94402830 I noticed that logview category was using quite a lot
of scribe quota, however we don't really use it much. The reason it was using a
lot of scribe quota is because we were logging all stderr messages to logview,
even "DEBUG" ones. That might also be the reason why we didn't use logview - it
was too spammy.

Let's make it possible to add an additional log level filter for logview, so
that we could log only e.g. warn log messages and above.

Reviewed By: Croohand

Differential Revision: D29456888

fbshipit-source-id: 8cc66773ca8d82b00c3337937f519f6140fc8c9d