Add redaction config to common config, don't use it yet
Summary:
This reads the config added on D29305462. It populates it into `CommonConfig` struct, and also adds it to `RepoFactory`, but doesn't yet use it anywhere, this will be done on the next diff.

There is a single behaviour change in this diff, which I believe should be harmless but is noted in the comments in case it isn't.

Reviewed By: markbt

Differential Revision: D29272581

fbshipit-source-id: 62cd7dc78478c1d8cb212eafdd789527ead50ef6